Remove SwipToDismiss of FlutterActivity
swiping from left to right on the watch screen
no longer dismisses the flutter activity. Now the back
button of the watch is the only way to exit the application and
instead of closing the application it will "minimize" the application
this way it can still collect workout metrics in the background.